add workout CRUD API, profile image support, and server sync columns
Backend (.NET API):
- Add Workout, WorkoutPlan, ScheduledWorkout models with full CRUD
  controllers, services, repositories, and DTOs
- Add EF Core migrations for workout session, workout entities, profile
  image URL column, and targetMuscleGroups rename
- Add ProfileImageUrl to User model, AuthResponseDto,
  UpdateProfileRequestDto, and UpdateProfileAsync signature

Flutter:
- Bump schema to v30: add server_id (nullable sync UUID) to all
  sync-able tables; add first_name, last_name, date_of_birth to
  user_table
- Update WorkoutExercise model to carry server_id
- Update Android build config

+++ FitTracker.Api/Models/WeigthTracking.cs
@@ -16,9 +16,9 @@ public class WeightTracking
     /// </summary>
     public DateTime Date { get; set; }
     /// <summary>
-    /// The weight of the user on the specified date. This property is of type decimal to allow for precise weight measurements, including fractional values.
+    /// The weight of the user on the specified date, stored as a double to match Flutter's real column type.
     /// </summary>
-    public decimal Weight { get; set; }
+    public double Weight { get; set; }
 
     /// <summary>An optional note attached to this weight entry.</summary>
     public string? Note { get; set; } = null;
